+++ c-blosc/blosc/blosc.c
@@ -883,8 +883,8 @@ static int32_t compute_blocksize(int32_t clevel, int32_t typesize,
       blocksize = MIN_BUFFERSIZE;
     }
   }
-  else if (nbytes >= L1*4) {
-    blocksize = L1 * 4;
+  else if (nbytes >= L1 * typesize) {
+    blocksize = L1 * typesize;
 
     /* For Zlib, increase the block sizes in a factor of 8 because it
        is meant for compression large blocks (it shows a big overhead
@@ -943,13 +943,6 @@ static int32_t compute_blocksize(int32_t clevel, int32_t typesize,
   /* blocksize must be a multiple of the typesize */
   if (blocksize > typesize) {
     blocksize = blocksize / typesize * typesize;
-  }
-
-  /* blocksize must not exceed (64 KB * typesize) in order to allow
-     BloscLZ to achieve better compression ratios (the ultimate reason
-     for this is that hash_log in BloscLZ cannot be larger than 15) */
-  if ((compressor == BLOSC_BLOSCLZ) && (blocksize / typesize) > 64*KB) {
-    blocksize = 64 * KB * typesize;
   }
 
   return blocksize;