+++ c-blosc/RELEASE_NOTES.rst
@@ -1,10 +1,26 @@
 ================================
- Release notes for c-blosc 1.4.2
+ Release notes for c-blosc 1.4.4
 ================================
 
 :Author: Francesc Alted
-:Contact: francesc@blosc.io
+:Contact: faltet@blosc.org
 :URL: http://www.blosc.org
+
+
+Changes from 1.4.3 to 1.4.4
+===========================
+
+* New computation of blocksize to be in sync with c-blosc 1.6.1
+
+* New parametrization of the hash table for blosclz (synced with c-blosc
+  1.6.1)
+
+
+Changes from 1.4.2 to 1.4.3
+===========================
+
+* Fixed a bug in blosclz that could potentially overwrite an area
+  beyond the output buffer.  See #113.
 
 
 Changes from 1.4.1 to 1.4.2
@@ -16,6 +32,7 @@ Changes from 1.4.1 to 1.4.2
 * Use multithreaded or single-threaded code depending on Blosc version.
   This is mainly useful for compiling HDF5 apps with non-included Blosc
   libraries.
+
 
 Changes from 1.4.0 to 1.4.1
 ===========================